test: Don't modify the filesystem during testing
Whether the tests passed was dependent on the order in which they ran,
and this was because some tests modified input data files (or rather,
the searcher did in the midst of the tests).  New files were also
created, which didn't cause tests to fail, but polluted the repository
(it'd be too easy to commit them).  This commit patches the methods
involved in modifying and writing files to be no-ops, leaving the
repository clean, and making it such that tests pass when run in any
order.